Fix apis and move existing tests
