+++ gamestonk_terminal/stocks/behavioural_analysis/ba_api.py
@@ -66,3 +66,5 @@ from gamestonk_terminal.common.behavioural_analysis.sentimentinvestor_view impor
 from gamestonk_terminal.stocks.behavioural_analysis.finnhub_view import (
     display_stock_price_headlines_sentiment as snews,
 )
+
+from .cramer_view import display_cramer_daily as cramer
