+++ tests/gamestonk_terminal/stocks/behavioural_analysis/csv/test_cramer_model/test_get_orders[False].csv
@@ -1,0 +1,24 @@
+,Date,Company,Symbol,Price,LastPrice,Change (%),Recommendation
+0,03/04,Allbirds ,BIRD,6.33,5.73,-10.47,Sell
+1,03/04,Bumble ,BMBL,18.08,16.66,-8.52,Sell
+2,03/04,Campbell Soup ,CPB,45.55,42.3,-7.68,Buy
+3,03/04,Dicks Sporting Goods ,DKS,109.71,101.58,-8.0,Buy
+4,03/04,DocuSign ,DOCU,101.38,91.98,-10.22,Sell
+5,03/04,Devon Energy ,DVN,59.57,59.48,-0.15,Buy
+6,03/04,eBay ,EBAY,55.66,53.31,-4.41,Sell
+7,03/04,Freeport-McMoRan ,FCX,50.11,46.67,-7.37,Buy
+8,03/04,Kohl's ,KSS,58.77,53.35,-10.16,Buy
+9,03/04,Macy's ,M,24.8,23.16,-7.08,Buy
+10,03/04,MongoDB ,MDB,306.11,281.74,-8.649999999999999,Sell
+11,03/04,On Holding ,ONON,21.97,20.83,-5.47,Sell
+12,03/04,Progress Software ,PRGS,44.28,42.48,-4.24,Sell
+13,03/04,The RealReal ,REAL,7.39,6.52,-13.34,Sell
+14,03/04,Rivian ,RIVN,47.39,42.21,-12.27,Sell
+15,03/04,Stitch Fix ,SFIX,11.17,11.01,-1.4500000000000002,Sell
+16,03/04,Virgin Galactic ,SPCE,7.84,7.42,-5.66,Sell
+17,03/04,Splunk ,SPLK,129.06,124.11,-3.9899999999999998,Buy
+18,03/04,AT&T ,T,23.87,23.0,-3.7800000000000002,Sell
+19,03/04,Trex ,TREX,74.68,68.44,-9.120000000000001,Sell
+20,03/04,Hostess Brands ,TWNK,22.22,21.06,-5.510000000000001,Buy
+21,03/04,Ulta Beauty ,ULTA,366.76,359.02,-2.16,Buy
+22,03/04,Wheels Up ,UP,3.13,3.39,7.670000000000001,Buy
