+++ tests/gamestonk_terminal/stocks/behavioural_analysis/test_cramer_model.py
@@ -2,7 +2,7 @@
 import pytest
 
 # IMPORTATION INTERNAL
-from gamestonk_terminal.stocks.discovery import cramer_view
+from gamestonk_terminal.stocks.behavioural_analysis import cramer_model
 
 
 @pytest.fixture(scope="module")
@@ -13,10 +13,10 @@ def vcr_config():
 
 
 @pytest.mark.vcr()
-@pytest.mark.record_stdout
 @pytest.mark.parametrize(
     "inverse",
     [True, False],
 )
-def test_get_orders(inverse):
-    cramer_view.display_cramer_daily(inverse=inverse)
+def test_get_orders(recorder, inverse):
+    recommendations = cramer_model.get_cramer_daily(inverse=inverse)
+    recorder.capture(recommendations)
