+++ tests/gamestonk_terminal/stocks/behavioural_analysis/txt/test_ba_controller/test_print_help.txt
@@ -30,4 +30,7 @@ Ticker: TSLA
 [SentimentInvestor]
     trend         most talked about tickers within the last hour
     hist          plot historical RHI and AHI data by hour
+[Jim Cramer]
+    jcdr          Jim Cramer's daily recommendations
+    jctr          Jim Cramer's recommendations by ticker
