--- tests/gamestonk_terminal/stocks/discovery/txt/test_cramer_view/test_get_orders[False].txt
@@ -1,22 +1,0 @@
-
-                       Company Symbol   Price  LastPrice  Change (%) Recommendation
-0                         AEO     AEO   21.33      19.34      -10.29            Buy
-1                       Cemig     CIG    2.61       2.67        2.25            Buy
-2               DigitalBridge    DBRG    7.40       7.37       -0.41           Sell
-3                         Dow     DOW   59.20      59.67        0.79            Buy
-4                  Ford Motor       F   18.10      17.60       -2.84            Buy
-5                 Foot Locker      FL   30.98      31.00        0.06           Sell
-6   Innovative Industrial Pro    IIPR  190.26     187.48       -1.48            Buy
-7         International Paper      IP   42.01      42.23        0.52            Buy
-8                   Nordstrom     JWN   26.93      25.68       -4.87            Buy
-9              Kontoor Brands     KTB   48.03      45.55       -5.44            Buy
-10                     Masimo    MASI  153.63     152.46       -0.77            Buy
-11             Morgan Stanley      MS   88.38      87.78       -0.68            Buy
-12      Norwegian Cruise Line    NCLH   19.06      18.56       -2.69            Buy
-13              Newell Brands     NWL   23.68      23.82        0.59            Buy
-14                       Okta    OKTA  182.70     167.98       -8.76            Buy
-15                     Pfizer     PFE   47.70      47.83        0.27            Buy
-16                  Snowflake    SNOW  264.69     224.02      -18.15            Buy
-17       Simon Property Group     SPG  138.77     139.41        0.46            Buy
-18                      Tesla    TSLA  879.89     839.29       -4.84           Sell
-
